Switch to using a composite action instead
There are few issues with a Docker container action.
For example, the path to the Azure Container Scan report is an absolute path in the workspace, but when using a Docker container action, the workspace is mounted in a '/github/workspace' folder, making it too complicated to reuse the path. This would require transforming the Container Scan output path so it can be used from within the container.